requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
Django==4.1.7
django-environ==0.9.0
environ==1.0
filelock==3.9.0
identify==2.5.18
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.0.4
psycopg2-binary==2.9.5
python-decouple==3.7
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.19.0

djangorestframework~=3.14.0
django-filter~=22.1
envparse~=0.2.0